Fix to not leak unneeded node types
