+++ lib/index.js
@@ -2,7 +2,7 @@
  * @typedef {import('unist').Node} Node
  * @typedef {import('unist').Position} Position
  * @typedef {import('unist').Point} Point
- * @typedef {import('./minurl.js').URL} URL
+ * @typedef {import('./minurl.shared.js').URL} URL
  *
  * @typedef {'ascii'|'utf8'|'utf-8'|'utf16le'|'ucs2'|'ucs-2'|'base64'|'latin1'|'binary'|'hex'} BufferEncoding
  *   Encodings supported by the buffer class.
