+++ lib/minurl.browser.js
@@ -1,5 +1,9 @@
 /// <reference lib="dom" />
 
+import {isUrl} from './minurl.shared.js'
+
+export {isUrl}
+
 // See: <https://github.com/nodejs/node/blob/fcf8ba4/lib/internal/url.js>
 
 /**
@@ -63,21 +67,4 @@ function getPathFromURLPosix(url) {
   }
 
   return decodeURIComponent(pathname)
-}
-
-// Note: this is copy/pasted from `minurl.js` because otherwise DOM/Node types
-// mix up.
-/**
- * @param {unknown} fileURLOrPath
- * @returns {fileURLOrPath is URL}
- */
-export function isUrl(fileURLOrPath) {
-  return (
-    fileURLOrPath !== null &&
-    typeof fileURLOrPath === 'object' &&
-    // @ts-expect-error: indexable.
-    fileURLOrPath.href &&
-    // @ts-expect-error: indexable.
-    fileURLOrPath.origin
-  )
 }
