+++ package.json
@@ -104,6 +104,10 @@
     "atLeast": 100,
     "detail": true,
     "strict": true,
-    "ignoreCatch": true
+    "ignoreCatch": true,
+    "#": "needed `any`s",
+    "ignoreFiles": [
+      "lib/minurl.shared.d.ts"
+    ]
   }
 }
